rename `Generators` to `Bolt3Derivation`
Note: `perCommitSecret`/`perCommitPoint` do not technically belong there,
in reality they are an implementation detail of `ChannelKeys`. But I think
it's simpler as-is.
