+++ src/commonMain/kotlin/fr/acinq/lightning/channel/ChannelData.kt
@@ -7,7 +7,7 @@ import fr.acinq.lightning.MilliSatoshi
 import fr.acinq.lightning.channel.Helpers.publishIfNeeded
 import fr.acinq.lightning.channel.Helpers.watchConfirmedIfNeeded
 import fr.acinq.lightning.channel.Helpers.watchSpentIfNeeded
-import fr.acinq.lightning.crypto.Generators
+import fr.acinq.lightning.crypto.Bolt3Derivation
 import fr.acinq.lightning.crypto.KeyManager
 import fr.acinq.lightning.transactions.Scripts
 import fr.acinq.lightning.transactions.Transactions.TransactionWithInputInfo.*
