+++ src/commonMain/kotlin/fr/acinq/lightning/channel/Helpers.kt
@@ -17,8 +17,8 @@ import fr.acinq.lightning.blockchain.fee.FeeratePerKw
 import fr.acinq.lightning.blockchain.fee.FeerateTolerance
 import fr.acinq.lightning.blockchain.fee.OnChainFeerates
 import fr.acinq.lightning.channel.Helpers.Closing.inputsAlreadySpent
-import fr.acinq.lightning.crypto.Generators.derive
-import fr.acinq.lightning.crypto.Generators.deriveRevocation
+import fr.acinq.lightning.crypto.Bolt3Derivation.derive
+import fr.acinq.lightning.crypto.Bolt3Derivation.deriveRevocation
 import fr.acinq.lightning.crypto.KeyManager
 import fr.acinq.lightning.crypto.ShaChain
 import fr.acinq.lightning.transactions.*
